cgsep almost done
Former-commit-id: 711d97bba79f980db3cf8fb2236e1a5b5eeb82bc

+++ mip/model.py
@@ -88,7 +88,7 @@ class LinExpr:
                  coeffs: List[float] = None,
                  const: float = 0,
                  sense: str = ""):
-        self.const: int = const
+        self.const: float = const
         self.expr: Dict[Var, float] = {}
         self.sense: str = sense
